+++ Dtos/PreparingMaterialDto.cs
@@ -7,14 +7,14 @@ public record PreparingMaterialDetailsDto(
     int MaterialId,
     string MaterialNumber,
     string MaterialDescription,
-    int PreparedQty,
+    decimal PreparedQty,
     string BUn
 );
 
 // -------------------- ใช้สำหรับ Insert / Update --------------------
 public record PreparingMaterialUpsertDto(
     int MaterialId,
-    int PreparedQty
+    decimal PreparedQty
 );
 
 // -------------------- สำหรับ List / Table View --------------------
@@ -22,5 +22,5 @@ public record PreparingMaterialListDto(
     int Id,
     int MaterialId,
     string MaterialNumber,
-    int PreparedQty
+    decimal PreparedQty
 );
