+++ Entities/WorkOrder.cs
@@ -17,7 +17,7 @@ public class WorkOrder
     public string OrderType { get; set; } = default!;       // Order Type
     public string Plant { get; set; } = default!;           // Plant
     public string Material { get; set; } = default!;        // Material information
-    public int Quantity { get; set; }                       // จำนวนที่จะผลิต
+    public decimal Quantity { get; set; }                       // จำนวนที่จะผลิต
     public string Unit { get; set; } = "PCE";               // หน่วยนับ (Default: PCE)
     public DateOnly? BasicFinishDate { get; set; }          // Basic Finish Date
     public string? DefaultLine { get; set; }                // Default Line from MES
